refactor: extract global search hook and user menu
